refactor(utils): rename util to utils and fix wrong require
